tws2json: fix off-by-one in final wait padding and update tests
turns out that the solution time should not include the time it would
take to execute the last move in the solution. the solution time is
equal to the time of the last input (or wait).

this doesn't affect most MS solutions, because they usually end with a
move onto the exit, so we don't need to add any padding. if chip reaches
the exit via an involuntary move like a force floor, then there will be
padding necessary.

it does affect Lynx solutions, because you have to wait for chip's
walking animation to complete before the game will acknowledge that
you've reached the exit, so most (all?) lynx solutions end with at least
4 frames of padding.